Crate mcc

Re-exports

pub extern crate target_lexicon;

pub use crate::assembling::assemble_and_link;

pub use crate::parsing::parse;

pub use crate::preprocessing::preprocess;

Modules

codegen

Compile Three Address Code to Assembly.

diagnostics

lowering

Lower from an Abstract Syntax Tree to Three Address Code.

render

types

Macros

codes

Declarative macro for defining error codes in a hierarchical structure.

Structs

Database

Files

Text

A reference-counted string.

Enums

CommandError

Traits

Db

Functions

assemble_and_link

Turn an assembly file into object code.

default_target

parse

Parse a C program into an abstract syntax tree.

preprocess

Run the C preprocessor on the given input.