mcc

Trait SerializeWithDatabase

pub trait SerializeWithDatabase {
    // Required method
    fn serialize_with_db<'a>(
        &'a self,
        db: &'a dyn Database,
    ) -> impl Serialize + 'a;
}

An adapter that lets you serialize a salsa type (e.g. from #[salsa::tracked] or #[salsa::input]) using serde.

You’ll most likely want to derive this using #[derive(mcc_macros::SerializeWithDatabase)].

Required Methods

fn serialize_with_db<'a>(&'a self, db: &'a dyn Database) -> impl Serialize + 'a

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.